modularised the JSON
