+++ src/commands/mod.rs
@@ -8,12 +8,6 @@ use crate::types::{FromRedisValue, NumericBehavior, RedisResult, ToRedisArgs, Re
 #[cfg(feature = "cluster")]
 use crate::cluster_pipeline::ClusterPipeline;
 
-#[cfg(feature = "json")]
-use serde::ser::Serialize;
-
-#[cfg(feature = "json")]
-use serde_json;
-
 #[cfg(feature = "geospatial")]
 use crate::geo;
 
@@ -22,6 +16,13 @@ use crate::streams;
 
 #[cfg(feature = "acl")]
 use crate::acl;
+
+#[macro_use]
+pub(crate) mod macros;
+
+/// Imports commands for RedisJSON, by importing `redis::json::Commands` and `redis::Commands`, you effectively merge them together
+#[cfg(feature = "json")]
+pub mod json;
 
 #[cfg(feature = "cluster")]
 pub(crate) fn is_readonly_cmd(cmd: &[u8]) -> bool {
@@ -56,285 +57,8 @@ pub(crate) fn is_readonly_cmd(cmd: &[u8]) -> bool {
     )
 }
 
-macro_rules! implement_commands {
-    (
-        $lifetime: lifetime
-        $(
-            $(#[$attr:meta])+
-            fn $name:ident<$($tyargs:ident : $ty:ident),*>(
-                $($argname:ident: $argty:ty),*) $body:block
-        )*
-    ) =>
-    (
-        /// Implements common redis commands for connection like objects.  This
-        /// allows you to send commands straight to a connection or client.  It
-        /// is also implemented for redis results of clients which makes for
-        /// very convenient access in some basic cases.
-        ///
-        /// This allows you to use nicer syntax for some common operations.
-        /// For instance this code:
-        ///
-        /// ```rust,no_run
-        /// # fn do_something() -> redis::RedisResult<()> {
-        /// let client = redis::Client::open("redis://127.0.0.1/")?;
-        /// let mut con = client.get_connection()?;
-        /// redis::cmd("SET").arg("my_key").arg(42).execute(&mut con);
-        /// assert_eq!(redis::cmd("GET").arg("my_key").query(&mut con), Ok(42));
-        /// # Ok(()) }
-        /// ```
-        ///
-        /// Will become this:
-        ///
-        /// ```rust,no_run
-        /// # fn do_something() -> redis::RedisResult<()> {
-        /// use redis::Commands;
-        /// let client = redis::Client::open("redis://127.0.0.1/")?;
-        /// let mut con = client.get_connection()?;
-        /// con.set("my_key", 42)?;
-        /// assert_eq!(con.get("my_key"), Ok(42));
-        /// # Ok(()) }
-        /// ```
-        pub trait Commands : ConnectionLike+Sized {
-            $(
-                $(#[$attr])*
-                #[inline]
-                #[allow(clippy::extra_unused_lifetimes, clippy::needless_lifetimes)]
-                fn $name<$lifetime, $($tyargs: $ty, )* RV: FromRedisValue>(
-                    &mut self $(, $argname: $argty)*) -> RedisResult<RV>
-                    { Cmd::$name($($argname),*).query(self) }
-            )*
-
-            /// Incrementally iterate the keys space.
-            #[inline]
-            fn scan<RV: FromRedisValue>(&mut self) -> RedisResult<Iter<'_, RV>> {
-                let mut c = cmd("SCAN");
-                c.cursor_arg(0);
-                c.iter(self)
-            }
-
-            /// Incrementally iterate the keys space for keys matching a pattern.
-            #[inline]
-            fn scan_match<P: ToRedisArgs, RV: FromRedisValue>(&mut self, pattern: P) -> RedisResult<Iter<'_, RV>> {
-                let mut c = cmd("SCAN");
-                c.cursor_arg(0).arg("MATCH").arg(pattern);
-                c.iter(self)
-            }
-
-            /// Incrementally iterate hash fields and associated values.
-            #[inline]
-            fn hscan<K: ToRedisArgs, RV: FromRedisValue>(&mut self, key: K) -> RedisResult<Iter<'_, RV>> {
-                let mut c = cmd("HSCAN");
-                c.arg(key).cursor_arg(0);
-                c.iter(self)
-            }
-
-            /// Incrementally iterate hash fields and associated values for
-            /// field names matching a pattern.
-            #[inline]
-            fn hscan_match<K: ToRedisArgs, P: ToRedisArgs, RV: FromRedisValue>
-                    (&mut self, key: K, pattern: P) -> RedisResult<Iter<'_, RV>> {
-                let mut c = cmd("HSCAN");
-                c.arg(key).cursor_arg(0).arg("MATCH").arg(pattern);
-                c.iter(self)
-            }
-
-            /// Incrementally iterate set elements.
-            #[inline]
-            fn sscan<K: ToRedisArgs, RV: FromRedisValue>(&mut self, key: K) -> RedisResult<Iter<'_, RV>> {
-                let mut c = cmd("SSCAN");
-                c.arg(key).cursor_arg(0);
-                c.iter(self)
-            }
-
-            /// Incrementally iterate set elements for elements matching a pattern.
-            #[inline]
-            fn sscan_match<K: ToRedisArgs, P: ToRedisArgs, RV: FromRedisValue>
-                    (&mut self, key: K, pattern: P) -> RedisResult<Iter<'_, RV>> {
-                let mut c = cmd("SSCAN");
-                c.arg(key).cursor_arg(0).arg("MATCH").arg(pattern);
-                c.iter(self)
-            }
-
-            /// Incrementally iterate sorted set elements.
-            #[inline]
-            fn zscan<K: ToRedisArgs, RV: FromRedisValue>(&mut self, key: K) -> RedisResult<Iter<'_, RV>> {
-                let mut c = cmd("ZSCAN");
-                c.arg(key).cursor_arg(0);
-                c.iter(self)
-            }
-
-            /// Incrementally iterate sorted set elements for elements matching a pattern.
-            #[inline]
-            fn zscan_match<K: ToRedisArgs, P: ToRedisArgs, RV: FromRedisValue>
-                    (&mut self, key: K, pattern: P) -> RedisResult<Iter<'_, RV>> {
-                let mut c = cmd("ZSCAN");
-                c.arg(key).cursor_arg(0).arg("MATCH").arg(pattern);
-                c.iter(self)
-            }
-        }
-
-        impl Cmd {
-            $(
-                $(#[$attr])*
-                #[allow(clippy::extra_unused_lifetimes, clippy::needless_lifetimes)]
-                pub fn $name<$lifetime, $($tyargs: $ty),*>($($argname: $argty),*) -> Self {
-                    ::std::mem::replace($body, Cmd::new())
-                }
-            )*
-        }
-
-        /// Implements common redis commands over asynchronous connections. This
-        /// allows you to send commands straight to a connection or client.
-        ///
-        /// This allows you to use nicer syntax for some common operations.
-        /// For instance this code:
-        ///
-        /// ```rust,no_run
-        /// use redis::AsyncCommands;
-        /// # async fn do_something() -> redis::RedisResult<()> {
-        /// let client = redis::Client::open("redis://127.0.0.1/")?;
-        /// let mut con = client.get_async_connection().await?;
-        /// redis::cmd("SET").arg("my_key").arg(42i32).query_async(&mut con).await?;
-        /// assert_eq!(redis::cmd("GET").arg("my_key").query_async(&mut con).await, Ok(42i32));
-        /// # Ok(()) }
-        /// ```
-        ///
-        /// Will become this:
-        ///
-        /// ```rust,no_run
-        /// use redis::AsyncCommands;
-        /// # async fn do_something() -> redis::RedisResult<()> {
-        /// use redis::Commands;
-        /// let client = redis::Client::open("redis://127.0.0.1/")?;
-        /// let mut con = client.get_async_connection().await?;
-        /// con.set("my_key", 42i32).await?;
-        /// assert_eq!(con.get("my_key").await, Ok(42i32));
-        /// # Ok(()) }
-        /// ```
-        #[cfg(feature = "aio")]
-        pub trait AsyncCommands : crate::aio::ConnectionLike + Send + Sized {
-            $(
-                $(#[$attr])*
-                #[inline]
-                #[allow(clippy::extra_unused_lifetimes, clippy::needless_lifetimes)]
-                fn $name<$lifetime, $($tyargs: $ty + Send + Sync + $lifetime,)* RV>(
-                    & $lifetime mut self
-                    $(, $argname: $argty)*
-                ) -> crate::types::RedisFuture<'a, RV>
-                where
-                    RV: FromRedisValue,
-                {
-                    Box::pin(async move { ($body).query_async(self).await })
-                }
-            )*
-
-            /// Incrementally iterate the keys space.
-            #[inline]
-            fn scan<RV: FromRedisValue>(&mut self) -> crate::types::RedisFuture<crate::cmd::AsyncIter<'_, RV>> {
-                let mut c = cmd("SCAN");
-                c.cursor_arg(0);
-                Box::pin(async move { c.iter_async(self).await })
-            }
-
-            /// Incrementally iterate set elements for elements matching a pattern.
-            #[inline]
-            fn scan_match<P: ToRedisArgs, RV: FromRedisValue>(&mut self, pattern: P) -> crate::types::RedisFuture<crate::cmd::AsyncIter<'_, RV>> {
-                let mut c = cmd("SCAN");
-                c.cursor_arg(0).arg("MATCH").arg(pattern);
-                Box::pin(async move { c.iter_async(self).await })
-            }
-
-            /// Incrementally iterate hash fields and associated values.
-            #[inline]
-            fn hscan<K: ToRedisArgs, RV: FromRedisValue>(&mut self, key: K) -> crate::types::RedisFuture<crate::cmd::AsyncIter<'_, RV>> {
-                let mut c = cmd("HSCAN");
-                c.arg(key).cursor_arg(0);
-                Box::pin(async move {c.iter_async(self).await })
-            }
-
-            /// Incrementally iterate hash fields and associated values for
-            /// field names matching a pattern.
-            #[inline]
-            fn hscan_match<K: ToRedisArgs, P: ToRedisArgs, RV: FromRedisValue>
-                    (&mut self, key: K, pattern: P) -> crate::types::RedisFuture<crate::cmd::AsyncIter<'_, RV>> {
-                let mut c = cmd("HSCAN");
-                c.arg(key).cursor_arg(0).arg("MATCH").arg(pattern);
-                Box::pin(async move {c.iter_async(self).await })
-            }
-
-            /// Incrementally iterate set elements.
-            #[inline]
-            fn sscan<K: ToRedisArgs, RV: FromRedisValue>(&mut self, key: K) -> crate::types::RedisFuture<crate::cmd::AsyncIter<'_, RV>> {
-                let mut c = cmd("SSCAN");
-                c.arg(key).cursor_arg(0);
-                Box::pin(async move {c.iter_async(self).await })
-            }
-
-            /// Incrementally iterate set elements for elements matching a pattern.
-            #[inline]
-            fn sscan_match<K: ToRedisArgs, P: ToRedisArgs, RV: FromRedisValue>
-                    (&mut self, key: K, pattern: P) -> crate::types::RedisFuture<crate::cmd::AsyncIter<'_, RV>> {
-                let mut c = cmd("SSCAN");
-                c.arg(key).cursor_arg(0).arg("MATCH").arg(pattern);
-                Box::pin(async move {c.iter_async(self).await })
-            }
-
-            /// Incrementally iterate sorted set elements.
-            #[inline]
-            fn zscan<K: ToRedisArgs, RV: FromRedisValue>(&mut self, key: K) -> crate::types::RedisFuture<crate::cmd::AsyncIter<'_, RV>> {
-                let mut c = cmd("ZSCAN");
-                c.arg(key).cursor_arg(0);
-                Box::pin(async move {c.iter_async(self).await })
-            }
-
-            /// Incrementally iterate sorted set elements for elements matching a pattern.
-            #[inline]
-            fn zscan_match<K: ToRedisArgs, P: ToRedisArgs, RV: FromRedisValue>
-                    (&mut self, key: K, pattern: P) -> crate::types::RedisFuture<crate::cmd::AsyncIter<'_, RV>> {
-                let mut c = cmd("ZSCAN");
-                c.arg(key).cursor_arg(0).arg("MATCH").arg(pattern);
-                Box::pin(async move {c.iter_async(self).await })
-            }
-        }
-
-        /// Implements common redis commands for pipelines.  Unlike the regular
-        /// commands trait, this returns the pipeline rather than a result
-        /// directly.  Other than that it works the same however.
-        impl Pipeline {
-            $(
-                $(#[$attr])*
-                #[inline]
-                #[allow(clippy::extra_unused_lifetimes, clippy::needless_lifetimes)]
-                pub fn $name<$lifetime, $($tyargs: $ty),*>(
-                    &mut self $(, $argname: $argty)*
-                ) -> &mut Self {
-                    self.add_command(::std::mem::replace($body, Cmd::new()))
-                }
-            )*
-        }
-
-        // Implements common redis commands for cluster pipelines.  Unlike the regular
-        // commands trait, this returns the cluster pipeline rather than a result
-        // directly.  Other than that it works the same however.
-        #[cfg(feature = "cluster")]
-        impl ClusterPipeline {
-            $(
-                $(#[$attr])*
-                #[inline]
-                #[allow(clippy::extra_unused_lifetimes, clippy::needless_lifetimes)]
-                pub fn $name<$lifetime, $($tyargs: $ty),*>(
-                    &mut self $(, $argname: $argty)*
-                ) -> &mut Self {
-                    self.add_command(::std::mem::replace($body, Cmd::new()))
-                }
-            )*
-        }
-    )
-}
-
 implement_commands! {
     'a
-    // most common operations
 
     /// Get the value of a key.  If key is a vec this becomes an `MGET`.
     fn get<K: ToRedisArgs>(key: K) {
@@ -409,138 +133,6 @@ implement_commands! {
     /// Set the expiration for a key as a UNIX timestamp.
     fn expire_at<K: ToRedisArgs>(key: K, ts: usize) {
         cmd("EXPIREAT").arg(key).arg(ts)
-    }
-
-    /// Append the JSON `value` to the array at `path` after the last element in it.
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_arrappend<K: ToRedisArgs, P: ToString, V: Serialize>(key: K, path: P, value: V) {
-        cmd("JSON.ARRAPPEND").arg(key).arg(path.to_string()).arg(serde_json::to_string(&value).unwrap())
-    }
-
-    /// Index array at `path`, returns first occurance of `value`
-    /// Pass `None` to `start_stop` to default it to `0, 0` (the redis default in this command)
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_arrindex<K: ToRedisArgs, P: ToString, V: Serialize>(key: K, path: P, value: V) {
-        cmd("JSON.ARRINDEX").arg(key).arg(path.to_string()).arg(serde_json::to_string(&value).unwrap())
-    }
-
-    /// Inserts the JSON `value` in the array at `path` before the `index` (shifts to the right)
-    /// `index` must be withing the array's range.
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_arrinsert<K: ToRedisArgs, P: ToString, V: Serialize>(key: K, path: P, index: i64, value: V) {
-        cmd("JSON.ARRINSERT").arg(key).arg(path.to_string()).arg(index).arg(serde_json::to_string(&value).unwrap())
-    }
-
-   /// Reports the length of the JSON Array at `path` in `key`.
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_arrlen<K: ToRedisArgs, P: ToString>(key: K, path: P) {
-        cmd("JSON.ARRLEN").arg(key).arg(path.to_string())
-    }
-
-    /// Removes and returns an element from the `index` in the array.
-    /// `index` defaults to `-1` (the end of the array)
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_arrpop<K: ToRedisArgs, P: ToString>(key: K, path: P, index: Option<i64>) {
-        cmd("JSON.ARRPOP").arg(key).arg(path.to_string()).arg(index.unwrap_or(-1))
-    }
-
-    /// Trims an array so that it contains only the specified inclusive range of elements.
-    ///
-    /// This command is extremely forgiving and using it with out-of-range indexes will not produce an error.
-    /// There are a few differences between how RedisJSON v2.0 and legacy versions handle out-of-range indexes.
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_arrtrim<K: ToRedisArgs, P: ToString>(key: K, path: P, start_stop: (i64, i64)) {
-        cmd("JSON.ARRTRIM").arg(key).arg(path.to_string()).arg(start_stop.0).arg(start_stop.1)
-    }
-
-    /// Clears container values (Arrays/Objects), and sets numeric values to 0.
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_clear<K: ToRedisArgs, P: ToString>(key: K, path: P) {
-        cmd("JSON.CLEAR").arg(key).arg(path.to_string())
-    }
-
-    /// Deletes a value at `path`.
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_del<K: ToRedisArgs, P: ToString>(key: K, path: P) {
-        cmd("JSON.DEL").arg(key).arg(path.to_string())
-    }
-
-    /// Returns the value at `path` in JSON serialized form.
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_get<K: ToRedisArgs, P: ToString>(key: K, path: P) {
-        cmd("JSON.GET").arg(key).arg(path.to_string())
-    }
-
-    /// Returns the values at `path` from multiple `key` arguments.
-    /// Returns `null` for nonexistent keys and nonexistent paths.
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_mget<K: ToRedisArgs, P: ToString>(keys: Vec<K>, path: P) {
-        cmd("JSON.MGET").arg(keys).arg(path.to_string())
-    }
-
-    /// Increments the number value stored at `path` by `number`.
-    #[cfg(feature = "json")]
-    fn json_numincrby<K: ToRedisArgs, P: ToString>(key: K, path: P, value: i64) {
-        cmd("JSON.NUMINCRBY").arg(key).arg(path.to_string()).arg(value)
-    }   
-
-    /// Returns the keys in the object that's referenced by `path`.
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_objkeys<K: ToRedisArgs, P: ToString>(key: K, path: P) {
-        cmd("JSON.OBJKEYS").arg(key).arg(path.to_string())
-    }
-
-    /// Reports the number of keys in the JSON Object at `path` in `key`.
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_objlen<K: ToRedisArgs, P: ToString>(key: K, path: P) {
-        cmd("JSON.OBJLEN").arg(key).arg(path.to_string())
-    }
-
-    /// Sets the JSON Value at `path` in `key`
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_set<K: ToRedisArgs, P: ToString, V: Serialize>(key: K, path: P, value: V) {
-        cmd("JSON.SET").arg(key).arg(path.to_string()).arg(serde_json::to_string(&value).unwrap())
-    }
-
-    /// Appends the `json-string` values to the string at `path`.
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_strappend<K: ToRedisArgs, P: ToString>(key: K, path: P, value: P) {
-        cmd("JSON.STRAPPEND").arg(key).arg(path.to_string()).arg(value.to_string())
-    }
-
-    /// Reports the length of the JSON String at `path` in `key`.
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_strlen<K: ToRedisArgs, P: ToString>(key: K, path: P) {
-        cmd("JSON.STRLEN").arg(key).arg(path.to_string())
-    }
-
-    /// Toggle a `boolean` value stored at `path`.
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_toggle<K: ToRedisArgs, P: ToString>(key: K, path: P) {
-        cmd("JSON.TOGGLE").arg(key).arg(path.to_string())
-    }
-
-    /// Reports the type of JSON value at `path`s.
-    #[cfg(feature = "json")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    fn json_type<K: ToRedisArgs, P: ToString>(key: K, path: P) {
-        cmd("JSON.TYPE").arg(key).arg(path.to_string())
     }
 
     /// Set a key's time to live in milliseconds.
